initial commit
We've begun writing some of our basic policies and mission statement. By "we," I mean me and the channel's two other members who I will not name because I don't know if they want to be called by pseudonyms. Now that I think of it, people will probably see their GitHub usernames. We might add a license to this repo, but I am not yet sure if that would be necessary or what license would be most appropriate.