mission statement and stuff
I added a mission statement to README.md, which is a work-in-progress, and rewrote that entire file to make that make more sense. I also added recommendations for CC-BY-SA 4.0 and Wikimedia Commons to the attribution section and changed the reasons we might not credit someone to be in an ordered list instead of an unordered one. I also rewrote the onjoin verification file with cleaner and more up-to-date info. And finally, I added a dedicated credits file.